backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
pymupdf==1.24.10
langgraph==0.2.34
langchain-core==0.3.15
python-multipart==0.0.12
aiosqlite==0.20.0
